# src/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateUserRequest:
    properties:
      is_admin:
        type: boolean
      password:
        type: string
      username:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_admin:
        type: boolean
      must_change_password:
        type: boolean
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.Webhook:
    properties:
      created_at:
        type: string
      events:
        items:
          $ref: '#/definitions/models.WebhookEvent'
        type: array
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      secret:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
  models.WebhookEvent:
    enum:
    - message.received
    - message.sent
    - message.failed
    type: string
    x-enum-varnames:
    - EventMessageReceived
    - EventMessageSent
    - EventMessageFailed
  models.WebhookRequest:
    properties:
      events:
        items:
          $ref: '#/definitions/models.WebhookEvent'
        type: array
      is_active:
        type: boolean
      name:
        type: string
      secret:
        type: string
      url:
        type: string
    type: object
...
host: localhost:5174
info:
  contact: {}
  description: REST API for sending and receiving SMS messages via a USB GSM modem.
  license:
    name: GPL-3.0
  title: SMS Gateway API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      description: Returns user accounts, newest first. Requires admin privileges.
      parameters:
      - description: Maximum users to return (max 500). Omit to return all.
        in: query
        name: limit
        type: integer
      - description: Users to skip. Only applied together with limit.
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Total users is returned in the X-Total-Count header
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Creates a new user account with the specified username, password,
        and admin status. Requires admin privileges.
      parameters:
      - description: User creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - Users
  /api/v1/users/{id}:
    delete:
      description: Permanently deletes a non-admin user account and its API keys.
        Messages sent with those keys are kept. Administrator accounts cannot be deleted.
        Requires admin privileges.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: user deleted'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - Users
  /api/v1/webhooks:
    get:
      description: Returns registered webhooks, newest first. Requires admin privileges.
      parameters:
      - description: Maximum webhooks to return (max 500). Omit to return all.
        in: query
        name: limit
        type: integer
      - description: Webhooks to skip. Only applied together with limit.
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Total webhooks is returned in the X-Total-Count header
          schema:
            items:
              $ref: '#/definitions/models.Webhook'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List webhooks
      tags:
      - Webhooks
    post:
      consumes:
      - application/json
      description: Registers a webhook notified on the selected message events. Leave
        secret empty to have one generated. Requires admin privileges.
      parameters:
      - description: Webhook to create
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.WebhookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Webhook'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create webhook
      tags:
      - Webhooks
  /api/v1/webhooks/{id}:
    delete:
      description: Permanently deletes a webhook. Requires admin privileges.
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: webhook deleted'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete webhook
      tags:
      - Webhooks
    put:
      consumes:
      - application/json
      description: Replaces a webhook's name, URL and events. An empty secret keeps
        the current one and an omitted is_active keeps the current state. Requires
        admin privileges.
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: string
      - description: New webhook settings
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.WebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Webhook'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update webhook
      tags:
      - Webhooks
securityDefinitions:
  ApiKeyAuth:
    description: API key for programmatic access
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: JWT token with Bearer prefix (e.g. "Bearer eyJhbG...")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"